+++ app/src/main/java/org/axonframework/samples/trader/app/api/order/CreateBuyOrderCommand.java
@@ -22,7 +22,7 @@ import org.axonframework.domain.AggregateIdentifier;
  */
 public class CreateBuyOrderCommand extends AbstractOrderCommand {
 
-    public CreateBuyOrderCommand(AggregateIdentifier portfolioId, AggregateIdentifier orderBookId, long tradeCount, int itemPrice) {
+    public CreateBuyOrderCommand(AggregateIdentifier portfolioId, AggregateIdentifier orderBookId, long tradeCount, long itemPrice) {
         super(portfolioId, orderBookId, tradeCount, itemPrice);
     }
 }
