+++ app/src/main/java/org/axonframework/samples/trader/app/api/order/SellOrderPlacedEvent.java
@@ -22,7 +22,7 @@ import org.axonframework.domain.AggregateIdentifier;
  */
 public class SellOrderPlacedEvent extends AbstractOrderPlacedEvent {
 
-    public SellOrderPlacedEvent(AggregateIdentifier orderId, long tradeCount, int itemPrice, AggregateIdentifier portfolioId) {
+    public SellOrderPlacedEvent(AggregateIdentifier orderId, long tradeCount, long itemPrice, AggregateIdentifier portfolioId) {
         super(orderId, tradeCount, itemPrice, portfolioId);
     }
 }
